admin/requirements.txt
# ***** MANI Requirements ***** #
django==3.2
python-dotenv==0.21.0
django-split-settings==1.1.0
pytest==7.1.1
psycopg2-binary==2.9.5
gunicorn==20.0.4
environ==1.0
docutils==0.19
django-ckeditor==6.5.1
django-quill-editor==0.1.40
factory_boy==3.2.1
halo==0.0.31
pyyaml==6.0
pillow==9.4.0
markdown==3.4.1
pygments==2.14.0
isort==5.11.4
pre-commit==2.21.0
autoflake==1.7.8
black==22.12.0
django-phonenumber-field[phonenumbers]==7.0.2

# ***** FLAKE 8 ***** #
flake8==5.0.0
flake8-blind-except==0.2.1
flake8-broken-line==0.6.0
flake8-bugbear==22.12.6
flake8-builtins==2.1.0
flake8-class-attributes-order==0.1.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-eradicate==1.4.0
flake8-functions==0.0.7
flake8-isort==6.0.0
flake8-mutable==1.2.0
flake8-print==5.0.0
flake8-pytest==1.4
flake8-pytest-style==1.6.0
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-variables-names==0.0.5
flake8-annotations==2.9.1
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-executable==2.1.2
flake8-fixme==1.1.1
flake8-future-import==0.4.7
flake8-pyi==22.11.0
flake8-todo==0.7
flake8-unused-arguments==0.0.12
